requirements.txt
allure-pytest==2.13.2
pre-commit==3.2.2
pytest==7.2.2
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.1
requests==2.28.2
structlog==23.1.0
PyHamcrest==2.0.4
selenium==3.14.0
